% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corr_data.R
\name{correlation_matrix}
\alias{correlation_matrix}
\title{Plot correlation plot correlation matrix}
\usage{
correlation_matrix(path)
}
\arguments{
\item{path}{Dataframe that includes columns with percentage of immune cells
and with bacterial counts or pathology scores.}

\item{correlation_data}{Merge the immune cell data with the other set of data to
calculate the correlation coefficient.}

\item{cor}{Method to calculate correlation coefficient values.Method can be changed to
"s" = Spearman or "p" = Pearson.}
}
\value{
A correlation matrix.
}
\description{
Plot correlation plot correlation matrix
}
