% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cellname.R
\name{fit_cellname}
\alias{fit_cellname}
\title{Fit cellname}
\usage{
fit_cellname(string, nwrap = 20)
}
\arguments{
\item{string}{Name of the cell type.}

\item{nwrap}{Number of cell type name character to wrap together.}
}
\value{
Split the name of the long cell type name into different sentences to fit into the graph.
}
\description{
Fit cellname
}
