% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{read_data}
\alias{read_data}
\title{Read data}
\usage{
read_data(path)
}
\arguments{
\item{path}{Path to the file, which you want to use for making graphs.}
}
\value{
Load the data in the required format to use in the pipeline.
}
\description{
Read data
}
