% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_function.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatterplot function}
\usage{
scatterPlot(data, method = "spearman")
}
\arguments{
\item{data}{Dataframe that includes a column with percentage of immune cells
and another column with bacterial counts or pathology scores.}

\item{method}{A method to test correlation between two objects.
Options include "spearman" and "pearson".}
}
\value{
Create a correlation plot comparing two parameters.
}
\description{
Scatterplot function
}
